Fix the DBPFEntry IsCompressed property behavior
The property now only indicates if the input data is actually
compressed, instead of also being used to signal if the data should be
or was compressed when saving.

This allows for things like reading back a modified property without
saving, and improves performance by only decompressing the data once.